Attempting to Implement Authentication

+++ lib/actions.ts
@@ -3,12 +3,64 @@
 import passwordModel from "./passwordModel";
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
-import { ZodPasswordSchemaServer } from "./zodDefinitions";
-import { passwordType } from "./typeDefinitions";
+import { ZodPasswordSchemaServer, ZodUserSchema } from "./zodDefinitions";
+import { authenticateUser, CreateUser, passwordType } from "./typeDefinitions";
 import mongoose from "mongoose";
+import { signIn } from "@/auth";
+import { AuthError } from "next-auth";
+import userModel from "./userModel";
 
+const AuthenticateUser = ZodUserSchema.omit({ name: true });
 const CreatePassword = ZodPasswordSchemaServer.omit({});
 const UpdatePassword = ZodPasswordSchemaServer.omit({ user_id: true });
+
+export async function authenticate(formData: authenticateUser) {
+  const parsedData = AuthenticateUser.safeParse(formData);
+
+  if (parsedData.success) {
+    console.log(
+      "Parsed Successfully_____________________________________________"
+    );
+    try {
+      await signIn("credentials", formData);
+    } catch (error) {
+      if (error instanceof AuthError) {
+        switch (error.type) {
+          case "CredentialsSignin":
+            return "Invalid credentials.";
+          default:
+            return "Something went wrong.";
+        }
+      }
+      throw error;
+    }
+  } else {
+    throw new Error("Invalid Credentials");
+  }
+}
+
+export const createUser = async (formData: CreateUser) => {
+  const parsedFormData = ZodUserSchema.safeParse(formData);
+
+  if (parsedFormData.success === false) {
+    console.log(parsedFormData.error);
+    throw new Error("Invalid Form Data");
+  } else {
+    try {
+      await userModel.create({
+        email_id: parsedFormData.data.email_id,
+        name: parsedFormData.data.name,
+        password: parsedFormData.data.password,
+      });
+    } catch (error) {
+      console.log(error);
+      throw new Error("Failed to create password");
+    }
+  }
+
+  revalidatePath("/dashboard");
+  redirect("/dashboard");
+};
 
 export const createPassword = async (formData: passwordType) => {
   const parsedFormData = CreatePassword.safeParse(formData);
@@ -73,7 +125,18 @@ export const updatePassword = async (
 };
 
 export const deletePassword = async (password_id: string) => {
-  await passwordModel.findByIdAndDelete(password_id);
+  const isValidId = mongoose.Types.ObjectId.isValid(password_id);
+
+  if (isValidId) {
+    try {
+      await passwordModel.findByIdAndDelete(password_id);
+    } catch (error) {
+      console.log(error);
+      throw new Error("Failed to Delete password");
+    }
+  } else {
+    throw new Error("Invalid Password ID.");
+  }
 
   revalidatePath("/dashboard/passwords");
   redirect("/dashboard/passwords");
